Break out Profile page into multiple components.
